+++ client/src/components/styles/Login.styled.ts
@@ -115,19 +115,6 @@ font-size: 0.7rem !important;
 }
 `;
 
-interface ErrorMessageProps {
-  showErr: number
-}
-
-export const ErrorMessage = styled.span<ErrorMessageProps>`
-display: ${({ showErr }) => (showErr ? 'inline' : 'none')};;
-width: 70%;
-text-align: start;
-color: #f02849;
-font-size: 0.55rem;
-font-family: Arial, Helvetica, sans-serif;
-`;
-
 export const SubmitButton = styled(StyledButton)`
 && {
   width: 70%;
